requirements.txt
numpy
tensorflow>=2.11.0
scikit-learn
matplotlib
pandas
opencv-python
seaborn
umap-learn
scikit-image
flask